notifier/requirements.txt
paho-mqtt==1.5.1
twilio==6.46.0
